+++ attributes/default.rb
@@ -6,3 +6,4 @@ default["envbuilder"]["data_bag"] = "envs"
 default["envbuilder"]["base_dbi"] = "development"
 default["envbuilder"]["joiner"] = "_"
 default["envbuilder"]["use_encrypted_data_bag"] = false
+default["envbuilder"]["file_permissions"] = "0644"