+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This file is used to list changes made in each version of envbuilder.
 
 * Default to using `node.chef_environment` if `node['ENV']`` is not set
   ([@prognostikos](https://github.com/prognostikos))
+* Add the ability to use an encrypted data bag
+  ([@prognostikos](https://github.com/prognostikos))
 
 ## 0.1.0: